pyproject.toml
[project]
name = "metube"
version = "0.1.0"
description = "Web GUI for youtube-dl (using yt-dlp) with playlist support"
requires-python = ">=3.13"
dependencies = [
    "aiohttp",
    "python-socketio>=5.0,<6.0",
    "yt-dlp[default,curl-cffi,deno]",
    "mutagen",
    "curl-cffi",
    "watchfiles",
    "python-telegram-bot>=22.6",
]

[dependency-groups]
dev = [
    "pylint",
]
